refactor(runtime): separate module execution and environments
